+++ l_hospital_mang/Data/Models/doctors_shifts.cs
@@ -2,10 +2,10 @@
 {
     public class doctors_shifts
     {
-        public int DoctorId { get; set; }
+        public long DoctorId { get; set; }
         public Doctors Doctor { get; set; }
 
-        public int ShiftId { get; set; }
+        public long ShiftId { get; set; }
         public Shifts Shift { get; set; }
     }
 }
